fix(path) Support absolute path with spaces or special characters
Absolute to relative path replacement was broken when the absolute file
path was conataing space characters. Space characters are subsitute with
their percent-encoding value in html `<img>` src attribute making the
regular expression pattern invalid in this case and resulting with a
preview having broken image references when the path to temp folder
contains spaces (project name or OS account name).

I don't known if those path can be at other places in the html document
where the path is not encoded. If this is the case we might need 2
regular pattern expression to replace.

In addition I have quoted the regexp pattern to support reserved regexp
characters in the path like `()` or `$`

[] Validated on Windows
[] Validated on Linux