enforce UTF-8 encoding of eclipse projects
By default, the encoding of Eclipse projects is that of the workspace.
And that in turn depends on the OS. Therefore importing the repository
on Windows leads to most projects being encoded as codepage 1252, not
UTF-8 (and German umlauts are garbled then).

Fix this by storing the encoding preference in each project. That way it
does not matter how the workspace encoding is configured.