make compatible with Eclipse Photon
The plugin does not use an explicit target platform, therefore it
compiles against the current IDE. That in turn makes adding versioned
dependencies a nightmare, since the version numbers depend on the IDE
version of the current developer. That's why the plugin is currently not
downwards compatible to older eclipse versions (some PR introduced high
version number dependencies).

Add an explicit target platform based on Eclipse photon. Describe in the
README how to use it. Fix the bad version numbers in the manifest to
become compatible with Eclipse Photon.