fix(edge): Use edge renderer
When using the internal preview on windows, the default renderer is
based on IE. Since a few eclipse version the SWT Browser supports Edge
on windows.

This PR suggest enabling the Edge renderer when available.
Using the modern renderer fix issues on windows where the svg images are
not display in the internal browser.